I declared the functions I changed a few things there's some problems in Classifier.cpp goodbye and kisses xoxo

+++ ChebyshevDistance.cpp
@@ -1,5 +1,5 @@
 #include "ChebyshevDistance.hpp"
 
-int ChebyshevDistance::dist(Point p1, Point p2) {
+double ChebyshevDistance::dist(Point p1, Point p2) {
     return abs(p1.getX() - p2.getX()) + abs(p1.getY() - p2.getY()) + abs(p1.getZ() - p2.getZ()) + abs(p1.getW() - p2.getW());
 }
